add(ergo): `ergo bump` command (releases, changelog)
Provides various utilities for managing versions, releases, and the
`CHANGELOG.md`.

Usage:

```shell
ergo bump changelog
ergo bump --help
```

Management of the `CHANGELOG.md` is powered by [`git-cliff`][^1]. The
approach is inspired by [Conventional Commits][^2] (for commit format
and parsing) and [Keep a Changelog][^3] for the format of the generated
content.

Note that we depart from the Conventional Commits specification. A small
collection of valid commit message prefixes is supported, and all are of
3-character length. The notion here is that fewer "hooks" to remember
ought to support uptake, and the invariant character length helps with
recall of what is available.

As of this commit, the available prefixes and configured mappings are:

- `^add` → "Added"
- `^del` → "Removed"
- `^fix` → "Fixed"
- `^mod` → "Changed"

[^1]: https://git-cliff.org
[^2]: https://www.conventionalcommits.org
[^3]: https://keepachangelog.com